requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
importlib-resources==1.5.0
paho-mqtt==1.5.0
pipenv==2018.11.26
psycopg2==2.8.5
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.19
virtualenv-clone==0.5.4
whitenoise==5.0.1
zipp==3.1.0
